methods for handing raw key directly to sqlite3_key() and sqlite_rekey()
This provides methods for handing the raw AES key data directly to
SQLCipher, skipping SQLCipher's key derivation process. This allows an app
to make their own key derivation process.  This is how we already handle
things with CacheWord, which uses 32 random bytes for the raw key, then
manages sending that raw key to multiple things that need to be unlocked
(SQLCipher-for-Android, IOCipher, RSA keys, etc).